update reSpec, update vocabulary files, docs
update reSpec: the respec css/script used in the DPV documentation has
been updated as per guidelines https://respec.org/docs/

update vocabulary file: the DPV files and modules are now generated by
the dpv-documentation repository which provides some guarentees as to
the correctness of the output. The vocabulary is mostly unchanged in
terms of terms and relationships. This is therefore purely a cosmetic
style change in terms of versioning.

docs: the HTML docs use the new reSpec style (see earlier message) and
are exported using the dpv-documentation scripts. They are therefore in
sync with the vocabularies/modules.

The dpv-documentation script used to generate these is at:
https://github.com/coolharsh55/dpv-documentation