dpv v0.4: dpv-gdpr,dpv-pd;dpv-owl,dpv-skos;Primer
DPV v0.4 makes important and backwards incompatible changes in terms of
how it expresses the concepts and relationships. It also moves the
namespace to w3id.org for all its documents. Most importantly, DPV is now
presented as an abstract 'data model' that is recorded using SKOS.
Specific serialisations are provided for use as an ontology in two
variants: first using SKOS+RDFS and the second using OWL2. In addition
to these, the personal data concepts have been moved to a separate
extension, i.e. DPV-PD, to prevent overlap and conflict with other
concepts still in DPV. A Primer has been added to provide an overview of
the DPV, its semantics, and concepts, and also to explain the above
changes in structure.

For more information, see Changelog.html