fix: missing concepts & processing context
v0.4 did not include concepts marked 'modified' due to an error in the
criteria for what is considered an 'accepted' concept. Fixed this in the
code for `002` rdf generation. The HTML documentation was missing the
inclusion of processing_context module, which has been fixed in the
`003` html generation code. See changelog for more info.