added proposed terms in docs ; deleted old files
The documentation now consists of an appendix listing the currently
proposed terms as applicable at the time of generation. Terms are stored
in a proposed.json file in each of the modules/folders. These are not
included as RDF given that they do not necessarily have the minimum
metadata expected from DPV terms.

Old files that were no longer needed or are superceded by other files
have been removed.