DPV v0.5, DPV-LEGAL v0.5, fixes
This commit updates the DPV to v0.5 (see changelog for specifics)
It adds the DPV-LEGAL extension, with an initial version 0.5 to match it
with the rest of DPV documents. The other extensions have also been
bumped to v0.5 for similar reasons.

The documentation generation now includes common code/text representing
the DPV family of documents to ensure the same links and text are
replicated across all documents, and a common set of references stored
in references.json.

This commit fixes the following issues:
- Fixes #34 regarding incorrect link
- Fix missing link to non-module concepts i.e. rights and risk not
- being defined in HTML
- Ensure consistency in links and text for DPV family of
- documents
- Add links to all documents (specs, extensions, guides)
- Fix incorrect IRIs in DPV-SKOS
- Fix incorrect or missing subclasses in DPV-SKOS