Added .owl (ManchesterSyntax) to DPV-OWL files
The DPV-OWL serialisations now include .owl ManchesterSyntax files in
addition to the other formats. The conversion is handled by `801`
script, and it will be run by `902` to auto-generate .owl files on every
update - same as other formats. The tool used for conversion is
onto-converter from https://github.com/sszuev/ont-converter/ In the
future, this can be replaced with a simpler OWLAPI wrapper.