DPV v0.8.1 - release candidate for v1.0
NOTE:
While this is a regular 0.x style release, it also acts as the v1.0
release candidate. This means if there are no major issues raised and
unable to be resolved, including missing parts, errors or bugs, or
requirements that must be added, then after 15-OCT-2022, this version
will be re-released as DPV v1.0 along with any further enhancements,
fixes, minor additions, or documentations that have been developed until
then.

This release includes:
- risks and rights have been added to DPV
- missing OWL/Manchester syntax files have been added/fixed
- consistent author affiliations
- funding acknowledgements
- MS-Office Excel (.xlsx) export format

In addition, see changelogs for DPV, DPV-GDPR, DPV-TECH regarding minor
changes to the concepts.