Add external concepts in using skos:scopeNote
External concepts, denoted using `prefix:term` in the source CSVs, are
not defined within the vocab, but are instead inserted with a
`skos:scopeNote` annotation (and appropriate Class/Property type). The
HTML documentation adds these with a section-prefixed ID (e.g. for
entities, it would be `entities-prefix:term`), with a reduced table
showing only the usage note and IRI of the term.

This change allows any property/concept to be added as a duplicate
within the various sections of the vocabulary in terms of providing
guidance on their use. The only restriction is that the Namespace be
known (added to the CSV) and that there can only be one term of the same
IRI in each module.