Adds source vocab for external terms
The use of external terms in HTML now adds a reference to the source
vocabulary as a hyperlink. For example, if the term was dct:title then a
row stating Vocabulary: <DCMI Metadata Terms link> is added.

This requires the reference to be present in references.json and the
corresponding prefix used in term be present in NAMESPACES variable
(this should already be happening through vocab_management.py)